backend/requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
confusable-homoglyphs==3.3.1
cookiecutter==2.6.0
Django==6.0.1
django-browser-reload==1.21.0
django-filter==25.2
django-registration==5.2.1
django-tailwind==4.4.2
djangorestframework==3.16.1
idna==3.11
Jinja2==3.1.6
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
psycopg2==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
